Shown page is now determined by the viewers
Before, the controller was responsible for determining which viewer
would see which raw pdf page.
It required multiple view-specific signals and connects.

The new system assigns "roles" to views that are propagated up to the
renderer for prefetching decisions.
Thus the controller only indicates which current presentation page is in
view.
Views are now responsible for determining the shown page from the
current page, using their role.

As a benefit, adding new roles only requires adding a new enum field and
new case in page_for_role ().