Merge timer within controller
Timer was split from Controller to try to separate concerns.
However it was tightly linked with it (signals both ways), so not the best idea.
And in order to introduce slide timing tracking, we need to manipulate both timing data AND slide position data.
This is easier if the timer is part of the controller.

Thus merged timer code inside the controller.
This removes complexity due to internal signals.
Also cleaned the API a bit, removed timer_start from slots.