Timings : usage, readme, fixed. Ready for 1.1
Added usage and readme doc for the timings feature.

Also fixed bug due to using page indexes instead of slide indexes.
This generated  buffer overflow when using transitions, as nb_page > nb_slides.
This could be improved by using typed indexes, but this is C++ here...  not worth the cost of changing the codebase.

With this feature pdftalk moves to version 1.1.
I am unlikely to touch the other features soon.